transcript/requirements.txt
SpeechRecognition==3.10.0
pydub==0.25.1
google-cloud-speech==2.21.1
openai==0.27.8
pyannote.audio
